+++ CHANGES.md
@@ -8,6 +8,7 @@ Version 0.3.2
 - added `each()`, `getChildren()`, `getItems()` to `flexgrid` component
 - added `bodyHeight` config to `modal` to control modal height
 - added `autocomplete`, `autocorrect`, `autocapitalize` config to `input`
+- added `onMouseDown`, `onMouseUp` events to `button`, `link`
 - fixed `getValues()`, `setValues()` work with nested components in `form`
 
 Version 0.3.1